Impl recursive referneced in static block
