+++ client-test/typescript/tests/runtime.test.ts
@@ -6,6 +6,7 @@ import {
   REFERENCED_LOCAL_VALUE_IN_STATIC_BLOCK, SELF_REFERENCED_LOCAL_VALUE_IN_STATIC_BLOCK, REFERENCED_IMPORTED_VALUE_IN_STATIC_BLOCK,
   REFERENCED_NESTED_VALUE_IN_STATIC_BLOCK, REFERENCED_STATIC_IMPORTED_VALUE_IN_STATIC_BLOCK,
   REFERENCED_PACKAGE_PRIVATE_VALUE_IN_STATIC_BLOCK, REFERENCED_SUPER_VALUE_IN_STATIC_BLOCK, SELECTED_SUPER_VALUE_IN_STATIC_BLOCK,
+  REFERENCED_STATIC_VALUE_IN_STATIC_BLOCK,
 } from '../src/index.java17';
 
 test('Constant values', () => {
@@ -32,4 +33,5 @@ test('Constant values', () => {
   expect(REFERENCED_PACKAGE_PRIVATE_VALUE_IN_STATIC_BLOCK).toBe(123);
   expect(REFERENCED_SUPER_VALUE_IN_STATIC_BLOCK).toBe(345);
   expect(SELECTED_SUPER_VALUE_IN_STATIC_BLOCK).toBe(345);
+  expect(REFERENCED_STATIC_VALUE_IN_STATIC_BLOCK).toBe(787);
 });
