+++ it/java17/src/test/java/online/sharedtype/it/ConstantsIntegrationTest.java
@@ -12,7 +12,7 @@ final class ConstantsIntegrationTest {
         ConstantNamespaceDef constantsDef = (ConstantNamespaceDef) deserializeTypeDef("$online.sharedtype.it.java8.MyConstants.ser");
         assertThat(constantsDef.simpleName()).isEqualTo("MyConstants");
         var components = constantsDef.components();
-        assertThat(components).hasSize(21);
+        assertThat(components).hasSize(22);
         assertThat(components).satisfiesExactly(
             component -> {
                 assertThat(component.name()).isEqualTo("FLOAT_VALUE");
@@ -97,6 +97,10 @@ final class ConstantsIntegrationTest {
             component -> {
                 assertThat(component.name()).isEqualTo("SELECTED_SUPER_VALUE_IN_STATIC_BLOCK");
                 assertThat(component.value()).isEqualTo(345L);
+            },
+            component -> {
+                assertThat(component.name()).isEqualTo("REFERENCED_STATIC_VALUE_IN_STATIC_BLOCK");
+                assertThat(component.value()).isEqualTo(787L);
             }
         );
     }
