+++ it/java8/src/main/java/online/sharedtype/it/java8/MyConstants.java
@@ -31,6 +31,7 @@ final class MyConstants extends IgnoredSuperClassB {
     static final long REFERENCED_PACKAGE_PRIVATE_VALUE_IN_STATIC_BLOCK;
     static final long REFERENCED_SUPER_VALUE_IN_STATIC_BLOCK;
     static final long SELECTED_SUPER_VALUE_IN_STATIC_BLOCK;
+    static final long REFERENCED_STATIC_VALUE_IN_STATIC_BLOCK;
     static {
         REFERENCED_VALUE_IN_STATIC_BLOCK = 112L;
         REFERENCED_LOCAL_VALUE_IN_STATIC_BLOCK = IGNORED_LOCAL_VALUE;
@@ -42,6 +43,7 @@ final class MyConstants extends IgnoredSuperClassB {
         REFERENCED_PACKAGE_PRIVATE_VALUE_IN_STATIC_BLOCK = JavaClass.SOME_LONG_VALUE;
         REFERENCED_SUPER_VALUE_IN_STATIC_BLOCK = LONG_VALUE_IN_SUPER;
         SELECTED_SUPER_VALUE_IN_STATIC_BLOCK = IgnoredSuperClassB.LONG_VALUE_IN_SUPER;
+        REFERENCED_STATIC_VALUE_IN_STATIC_BLOCK = OtherConstants.Inner1.INNER_LONG_VALUE_IN_STATIC_BLOCK;
     }
 }
