+++ it/java8/src/main/java/online/sharedtype/it/java8/TempClass.java
@@ -4,9 +4,6 @@ import online.sharedtype.SharedType;
 import online.sharedtype.it.java8.other.OtherConstants;
 
 @SharedType(includes = SharedType.ComponentType.CONSTANTS)
-public class TempClass extends IgnoredSuperClassB{
-    static final long REFERENCED_VALUE;
-    static {
-        REFERENCED_VALUE = OtherConstants.Inner1.Inner2.INNER_LONG_VALUE;
-    }
+public class TempClass extends IgnoredSuperClassB {
+    //    static final EnumGalaxy REFERENCED_VALUE = EnumGalaxy.MilkyWay;
 }
