+++ it/java8/src/main/java/online/sharedtype/it/java8/other/OtherConstants.java
@@ -5,6 +5,11 @@ public final class OtherConstants {
 
     public static final class Inner1 {
         public static final long STATIC_IMPORTED_VALUE = 999L;
+        public static final long INNER_LONG_VALUE_IN_STATIC_BLOCK;
+        static {
+            INNER_LONG_VALUE_IN_STATIC_BLOCK = 787L;
+        }
+
         public static final class Inner2 {
             public static final long INNER_LONG_VALUE = 777L;
         }
